fix: 修复环境变量加载、文件路径验证及异步文档解析问题
- 显式加载 .env 文件以确保环境变量正确初始化
- 修复 Windows 路径验证和文件访问权限检查
- 将文档解析改为异步操作以避免阻塞，并设置 HF 镜像解决连接问题
- 增加图标服务、文件上传和轮廓生成的错误处理和日志记录
- 更新依赖版本并修正 Next.js 代理配置
- 添加项目工作流和数据结构文档

+++ servers/fastapi/models/pptx_models.py
@@ -128,7 +128,7 @@ class PptxAutoShapeBoxModel(PptxShapeModel):
     shadow: Optional[PptxShadowModel] = None
     position: PptxPositionModel
     text_wrap: bool = True
-    border_radius: Optional[int] = None
+    border_radius: Optional[float] = None
     paragraphs: Optional[List[PptxParagraphModel]] = None
